+++ docker/bootstrap/dgoss/test-with-pre-run/pre_run.sh
@@ -5,3 +5,6 @@ echo "** this is a pre-scrip for run.sh that can be edit with configmap"
 grep max_hints_delivery_threads /etc/cassandra/cassandra.yaml
 sed -i 's/max_hints_delivery_threads: 2/max_hints_delivery_threads: 8/' /etc/cassandra/cassandra.yaml
 grep max_hints_delivery_threads /etc/cassandra/cassandra.yaml
+
+sed -i 's/authenticator: AllowAllAuthenticator/authenticator: PasswordAuthenticator/' /etc/cassandra/cassandra.yaml
+sed -i 's/authorizer: AllowAllAuthorizer/authorizer: CassandraAuthorizer/' /etc/cassandra/cassandra.yaml