+++ Changelog.md
@@ -4,8 +4,8 @@
 
 ## 0.4.0
 
-- Uses of initContainerImage and bootstrapContainerImage to be able to adapt to official cassandra image.
-- Use of ReadOnly Container : `Spec.ReadOnlyRootFilesystem` default true
+- initContainerImage and bootstrapContainerImage used to adapt to official cassandra image.
+- ReadOnly Container : `Spec.ReadOnlyRootFilesystem` default true
 
 ## 0.3.3